requirements.txt
# Core Web/API Framework
fastapi
uvicorn

# Database and ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
geoalchemy2

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Geospatial Processing
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0
fiona>=1.9.0

# Machine Learning / Forecasting
prophet
scikit-learn>=1.3.0

# AI Analysis Dependencies for Government Data Processing
# Core machine learning and data analysis libraries
scipy>=1.9.0

# HTTP requests for API testing
requests>=2.28.0

# Core AI Dependencies (Required)
scikit-learn>=1.3.0
scipy>=1.9.0
numpy>=1.24.0
pandas>=2.0.0

# Text processing and NLP
nltk>=3.8
spacy>=3.4.0

# Data validation and quality
great-expectations>=0.17.0
pandera>=0.14.0

# Statistical analysis
statsmodels>=0.13.0

# Feature extraction
feature-engine>=1.5.0

# Model persistence
joblib>=1.2.0

# Data Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# Graph Visualization
pydot>=1.4.0
graphviz>=0.20.0

# JSON processing
jsonschema>=4.0.0

# XML processing
lxml>=4.9.0
xmltodict>=0.12.0

# Regular expressions and pattern matching
regex>=2022.8.0

# Date and time processing
python-dateutil>=2.8.0
pytz>=2022.1

# Configuration management
pyyaml>=6.0

# Logging and monitoring
structlog>=23.0.0

# Performance optimization
numba>=0.56.0

# Parallel processing
multiprocessing-logging>=0.3.0

# Memory optimization
psutil>=5.9.0
memory-profiler>=0.60.0

# Caching
diskcache>=5.2.0

# API rate limiting
ratelimit>=2.2.0

# Data compression
zstandard>=0.20.0

# Progress tracking
tqdm>=4.65.0

# Utilities and Performance
python-multipart

# Optional: Advanced Geospatial Operations
rtree>=1.0.0
geopy>=2.3.0

# Optional: Fuzzy String Matching
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0

# Optional: Additional Data Analysis
jupyter>=1.0.0
ipykernel>=6.25.0

# Authentication and Security
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0
fastapi[security]

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
black>=22.0.0
flake8>=5.0.0
mypy>=0.991 

